requirements.txt
wheel
fastapi
uvicorn
openai
bs4
requests
pandas
psycopg
python-dotenv
scipy
slowapi
tiktoken
lxml
asyncpg
sqlalchemy
sqlmodel